Use LayerPayload instead of Payload.
According to the gopacket documentation, the LayerPayload is just the
payload of the layer, minus the headers. This should help us get a more
accurate byte count of the requests and responses going to and from the
database.

If you inspect a layer's Payload you'll see lots of terminating 0s and
1s, indicative of headers. If you inspect the LayerPayload, you'll
notice it looks like all be legitimate data.

https://godoc.org/github.com/google/gopacket#Layer