Loop over responses only once, loop over metrics every time.
If I'm understanding the current code correctly, we loop over all (or at
least most) of the responses for every single QueryMetric, in order to
associate responses with metrics. This leads to O(n^2) complexity, and
causes it to take a long time to process more and more packets.

This switches over to running through the responses once, and iterating
over the metrics to find the associated Seq. While I think this is still
actually O(n^2) complexity, it's definitely less intensive.

Benchmarks show similar performance in the double digit MB, for a 600MB
run there was an average 15s improvement, and I didn't ever have the
patience to run it for anything in the GB.

Results are also identical.